+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**0.1.0** - TBD
+
+- export the keys of plain objects as named exports, in addition to exporting
+  the object itself as "default", "default_1" etc.
+
 **0.0.2** - 2020-04-26
 
 - add missing documentation